Possible refactoring of export functionality
